feat: Implement single file distribution with Costura.Fody
This commit introduces changes that enable the application to be distributed as a single executable file. This was achieved by using the Costura.Fody library, which embeds all the dependencies into the main assembly.

With this update, our application's deployment process is simplified and the user experience is improved, as users can now run the application without needing to handle multiple files or dependencies.

Please note, this is a significant change in the way our application is packaged and distributed. All tests have passed, but further testing is recommended to ensure there are no unforeseen issues.